pyproject.toml
[project]
name = "whisper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "openai-whisper>=20250625",
]
